finished refactor for productroutes with non-codesplitting.
